fix: remove pleasantries
This is an opinion, happy for it to be ignored, but I'd suggest omitting the word "please" in copy as a soft rule. Please is used widely, but oftentimes it's just an extra word that doesn't add anything. Does a reader _really_ ever think "Hmm, this user interface is being so rude to me" when a plain instruction is given? I think this suggestion is inline with Material Design's recommendations _Click continue_ vs _Please click continue_ etc